Add discount property to db models and refactor into separate files
