+++ Reservation/package.json
@@ -5,10 +5,11 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "makeDB": "mysql -u root -p < ./database/schema.sql",
+    "makeDB": "mysql -u root -p < ./database/makeDB.sql",
     "seed": "node ./database/seed.js",
     "build": "webpack -d -w",
-    "date": "node ./database/mockGenerator.js"
+    "start": "nodemon ./server/index.js",
+    "check": "node ./database/mockGenerator.js"
   },
   "author": "",
   "license": "ISC",
@@ -31,6 +32,7 @@
     "webpack-cli": "^3.3.9"
   },
   "dependencies": {
+    "express": "^4.17.1",
     "moment": "^2.24.0",
     "mysql": "^2.17.1",
     "mysql2": "^1.7.0",
